Initial commit for letter frequency counter
This commit adds a C++ program that prompts the user for a string input, counts the frequency of each letter in the input, and then iterates through a map to print out the frequency of each letter.

The program consists of three functions: promptForInput, countFrequency, and iterateThroughMap. The main function calls these functions in sequence to run the program.

No additional files or dependencies are needed for this program.

Author: [Jack David Mullins]
Date: [25th March 2023]